requirements.txt
mangum
strawberry-graphql[fastapi]
fastapi
uvicorn
pydantic==1.10
